Add Flow extension to present a native UIAlertController alert by providing a title and message. Update PresentError extension to call the presentAlert() extension.
